Initial work on indeterminate checkbox
