+++ web/html/src/components/input/check/Check.module.scss
@@ -1,0 +1,15 @@
+@import "branding/css/variables.scss";
+
+:global(.old-theme),
+:global(.new-theme) {
+  .check {
+    display: flex;
+    gap: 6px;
+    white-space: nowrap;
+    align-items: center;
+
+    input {
+      --bs-border-width: 1px;
+    }
+  }
+}
