+++ web/html/src/components/input/check/Check.tsx
@@ -1,63 +1,33 @@
-import * as React from "react";
+import { useEffect, useRef } from "react";
 
-import { ControlledInput } from "../ControlledInput";
-import { FormContext } from "../form/Form";
-import { InputBase, InputBaseProps } from "../InputBase";
+import styles from "./Check.module.scss";
 
-type Props = Omit<InputBaseProps, "label"> & {
-  /** CSS class for the <input> element */
-  inputClass?: string;
-
-  /** name of the field to map in the form model */
-  name: string;
-
-  /** for historical reasons, label handles differently here  */
-  label: React.ReactNode;
+export type CheckProps = Omit<React.InputHTMLAttributes<HTMLInputElement>, "value" | "onChange"> & {
+  indeterminate?: boolean;
+  onChange?: (value: boolean) => void;
 };
 
-export function Check(props: Props) {
-  const { label, inputClass, ...propsToPass } = props;
-  const formContext = React.useContext(FormContext);
+export const Check = (props: CheckProps) => {
+  const ref = useRef<HTMLInputElement>(null);
+
+  const { indeterminate, ...rest } = props;
+
+  useEffect(() => {
+    if (typeof indeterminate !== "undefined" && ref.current) {
+      ref.current.indeterminate = indeterminate;
+    }
+  }, [indeterminate]);
 
   return (
-    <InputBase {...propsToPass}>
-      {({ setValue, onBlur }) => {
-        const setChecked = (event: any) => {
-          setValue(event.target.name, event.target.checked);
-        };
-        const fieldValue = (formContext.model || {})[props.name] || props.defaultValue || "";
-        return (
-          <div className="checkbox">
-            <label htmlFor={props.name}>
-              <ControlledInput
-                id={props.name}
-                className={inputClass}
-                name={props.name}
-                type="checkbox"
-                checked={fieldValue}
-                onChange={setChecked}
-                onBlur={onBlur}
-                disabled={props.disabled}
-                title={props.title}
-              />
-              <span>{label}</span>
-            </label>
-          </div>
-        );
-      }}
-    </InputBase>
+    <label className={styles.check}>
+      <input
+        {...rest}
+        className={`form-check-input ${props.className || ""}`}
+        checked={props.checked}
+        onChange={(event) => props.onChange?.(event.target.checked)}
+        type="checkbox"
+        ref={ref}
+      />
+    </label>
   );
-}
-
-Check.defaultProps = {
-  inputClass: undefined,
-  defaultValue: undefined,
-  label: undefined,
-  hint: undefined,
-  labelClass: undefined,
-  divClass: undefined,
-  required: false,
-  disabled: false,
-  invalidHint: undefined,
-  onChange: undefined,
 };
