+++ web/html/src/manager/maintenance/edit/calendar-edit.tsx
@@ -6,7 +6,7 @@ import validator from "validator";
 import { Button } from "components/buttons";
 import { DangerDialog } from "components/dialog/LegacyDangerDialog";
 import { ModalButton } from "components/dialog/ModalButton";
-import { Check } from "components/input/check/Check";
+import { DEPRECATED_Check } from "components/input/check/DEPRECATED_Check";
 import { Form } from "components/input/form/Form";
 import { Text } from "components/input/text/Text";
 import { MessageType, Utils as MessagesUtils } from "components/messages/messages";
@@ -128,7 +128,7 @@ const MaintenanceCalendarEdit = forwardRef((props: CalendarEditProps, ref) => {
         disabled={props.isEdit}
       />
       {props.isEdit && !props.calendar?.url && (
-        <Check
+        <DEPRECATED_Check
           name="strategy"
           label={<b>{t("Cancel affected actions")}</b>}
           divClass="col-md-6 col-md-offset-3 offset-md-3"
@@ -184,7 +184,7 @@ const MaintenanceCalendarEdit = forwardRef((props: CalendarEditProps, ref) => {
                   <div>
                     <div>{t("Refreshing the calendar will reschedule all affected actions.")}</div>
                     <div>{t("Confirm if you want to proceed.")}</div>
-                    <Check name="strategy" label={<b>{t("Cancel affected actions?")}</b>} divClass="col-md-6" />
+                    <DEPRECATED_Check name="strategy" label={<b>{t("Cancel affected actions?")}</b>} divClass="col-md-6" />
                   </div>
                 }
                 onConfirm={() => onConfirmRefresh()}
